Add ASAN builds
It compiles some of Ruby's key dependencies with ASAN enabled, ships
those dynamic libraries to a location in Ruby's prefix (which is part of
the tarball deployed with setup-ruby). -Wl,-rpath is used to make sure
that these ASANified libs are used in preference to any
distribution-provided libs at runtime.

Only Ubuntu 24.04 is supported, because the version of clang in 22.04 is
too old. MacOS is also not yet supported.