Add 3.4-asan builds
**What does this PR do?**

This PR introduces a new "3.4-asan" build, based on the existing
asan builds, but just pointed at whatever's the latest tag in the
3.4 series.

**Motivation:**

The intention of "3.4-stable" is to provide the latest up-to-date
stable Ruby, so that we can reliably use it as a breaking CI step.

As discussed in https://github.com/ruby/setup-ruby/issues/682, the
current ruby-asan builds are a bit of a "sharp edge" when used in CI
because they may break due to changes that are completely unrelated
to asan.

Building asan rubies is a bit awkward still, as e.g. ruby-build and
other version managers don't have support for it, and it requires
very modern versions of specific system tools (e.g. clang).

**Additional Notes:**

After some back-and-forth, the changes are reasonably minimal.

In particular, I decided to not touch the logic that determines
weather there's a more recent commit to build or not. This does
mean that if ruby master sees no commits, but there's a new 3.4
stable release, this won't be picked up immediately; and it
also means that if there's a new master commit and no change
to the 3.4 branch we still rebuild 3.4-asan.

My thinking is that the extra complexity to individually take
care of the caching for both branches is not worth the trouble vs
doing some extra rebuilds for 3.4-asan.

Let me know if you're not convinced, and I can change that.

**How to test the change?**

I've built this in the downstream fork, and manually downloaded
the resulting Ruby and it seems to be in good shape and with
asan working fine.

* Successful run:
  https://github.com/DataDog/ruby-dev-builder/actions/runs/12827351740
* Resulting builds:
  https://github.com/DataDog/ruby-dev-builder/releases/tag/v20250117.103455